requirements.txt
streamlit
openai-whisper
deep-translator
